ETL module structure (flat split, move-only)
